Add configuration option for custom module paths.
Even though I don't think it's a good idea to use this, I'm not going to
convince everyone and it's been requested by several people for over a year.

It was proposed we use [`moduleResolve`][tweet] as the key. I don't think it's
descriptive enough and is too similar to webpack's
[`modulesDirectories`][webpack] and [resolve's `moduleDirectory`][resolve].
Those options also don't really provide the correct behavior for this.

I chose `moduleRoots` because the behavior for this feature is most similar to
[webpack's `resolve.root`][webpack.root]. Webpack says you must use absolute
paths, but since our configuration is in JSON you must use paths relative to
`package.json`

Fixes #17
Fixes #27

[tweet]: https://twitter.com/nodkz/status/771735917203689472
[webpack]: http://webpack.github.io/docs/configuration.html#resolve-modulesdirectories
[resolve]: https://www.npmjs.com/package/resolve
[webpack.root]: http://webpack.github.io/docs/configuration.html#resolve-root